requirements.txt
python-telegram-bot==20.*
mariadb
selenium
requests
beautifulsoup4
python-dotenv
